Add get gpx trace metadata command

+++ trace.pl
@@ -24,6 +24,13 @@ if ($ARGV[0] eq "create")
     }
 }
 
+if (($ARGV[0] eq "get") && (scalar(@ARGV) == 2))
+{
+    my $content = Trace::get($ARGV[1]);
+    print $content;
+    exit;
+}
+
 if (($ARGV[0] eq "delete") && (scalar(@ARGV) == 2))
 {
     Trace::delete($ARGV[1]);
@@ -31,8 +38,9 @@ if (($ARGV[0] eq "delete") && (scalar(@ARGV) == 2))
 }
 
 print <<EOF;
-Usage: 
+Usage:
   $0 create <filename> <options>    upload new gpx trace
+  $0 get <id>                       load and print trace metadata XML
   $0 delete <id>                    delete trace
 
 options:
